Update index.py
Modified the condition to check for the presence of special characters.
Originally, the code checked if a character is not lowercase, uppercase, or a space,
but we needed to also check for special characters. Therefore, we added the condition
'elif not char.isalnum()' to identify special characters in the password.